+++ lib/trip/trip_screen.dart
@@ -1,4 +1,4 @@
-import 'package:blueboard/widgets/drawer/app_drawer.dart';
+import 'package:blueboard/widgets/drawer/drawer.dart';
 import 'package:flutter/material.dart';
 
 class TripPage extends StatefulWidget {
